Support independent, non-global Events objects
Adds a public New() method that creates a new object with its own set of
event channels. The created object is part of a new Events interface
with the public API: Announce, Listen and Notify.

Remains backwards-compatible by instantiating a global Events object and
forwarding the Announce, Listen and Notify methods to that object.